pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "streetlevel"
version = "0.8.1"
license = "MIT"
authors = [
  { name="sk-zk", email="sk-zk@users.noreply.github.com" },
]
description = "Download panoramas and metadata from Street View and others"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
    "Pillow",
    "aiohttp",
    "cython",
    "numpy",
    "protobuf > 4",
    "pycryptodome",
    "pyequilib",
    "pyfrpc",
    "pyproj",
    "pytest",
    "pytest-asyncio",
    "requests",
    "scipy",
]

[project.urls]
"Homepage" = "https://github.com/sk-zk/streetlevel"
"Bug Tracker" = "https://github.com/sk-zk/streetlevel/issues"
